# Improve vertex upload performance significantly on iOS