--- a/osu.Framework/Graphics/Veldrid/Batches/VeldridLinearBatch.cs
+++ b/osu.Framework/Graphics/Veldrid/Batches/VeldridLinearBatch.cs
@@ -3,7 +3,6 @@
 
 using System;
 using osu.Framework.Graphics.Rendering.Vertices;
-using osu.Framework.Graphics.Veldrid.Buffers;
 using Veldrid;
 
 namespace osu.Framework.Graphics.Veldrid.Batches
@@ -11,14 +10,9 @@ namespace osu.Framework.Graphics.Veldrid.Batches
     internal class VeldridLinearBatch<T> : VeldridVertexBatch<T>
         where T : unmanaged, IEquatable<T>, IVertex
     {
-        private readonly PrimitiveTopology type;
-
         public VeldridLinearBatch(VeldridRenderer renderer, int size, PrimitiveTopology type)
-            : base(renderer, size)
+            : base(renderer, size, type)
         {
-            this.type = type;
         }
-
-        protected override VeldridVertexBuffer<T> CreateVertexBuffer(VeldridRenderer renderer) => new VeldridLinearBuffer<T>(renderer, Size, type);
     }
 }