--- a/osu.Framework/Graphics/Veldrid/Batches/VeldridQuadBatch.cs
+++ b/osu.Framework/Graphics/Veldrid/Batches/VeldridQuadBatch.cs
@@ -2,21 +2,20 @@
 // See the LICENCE file in the repository root for full licence text.
 
 using System;
+using osu.Framework.Graphics.Rendering;
 using osu.Framework.Graphics.Rendering.Vertices;
-using osu.Framework.Graphics.Veldrid.Buffers;
+using PrimitiveTopology = Veldrid.PrimitiveTopology;
 
 namespace osu.Framework.Graphics.Veldrid.Batches
 {
     internal class VeldridQuadBatch<T> : VeldridVertexBatch<T>
         where T : unmanaged, IEquatable<T>, IVertex
     {
-        public VeldridQuadBatch(VeldridRenderer renderer, int size)
-            : base(renderer, size)
+        public VeldridQuadBatch(VeldridRenderer renderer, int quads)
+            : base(renderer, quads * IRenderer.VERTICES_PER_QUAD, PrimitiveTopology.TriangleList)
         {
-            if (size > VeldridQuadBuffer<T>.MAX_QUADS)
-                throw new OverflowException($"Attempted to initialise a {nameof(VeldridQuadBuffer<T>)} with more than {nameof(VeldridQuadBuffer<T>)}.{nameof(VeldridQuadBuffer<T>.MAX_QUADS)} quads ({VeldridQuadBuffer<T>.MAX_QUADS}).");
+            if (quads > IRenderer.MAX_QUADS)
+                throw new OverflowException($"Attempted to initialise a {nameof(VeldridQuadBatch<T>)} with more than {nameof(IRenderer)}.{nameof(IRenderer.MAX_QUADS)} quads ({IRenderer.MAX_QUADS}).");
         }
-
-        protected override VeldridVertexBuffer<T> CreateVertexBuffer(VeldridRenderer renderer) => new VeldridQuadBuffer<T>(renderer, Size);
     }
 }