- Closes https://github.com/ppy/osu-framework/issues/5920

With the new performance test scenes we have, VBO uploads have shown to be quite poor on iOS, even when building in AOT.

After multiple days of investigation (interrupted by regressions in our rendering systems), it turns out the intermediate vertex storage allocated in `VeldridVertexBuffer` have a significant hit on performance. In an isolated environment, there's +7ms aggregate CPU overhead when uploading to an intermediate vertex storage followed by uploading to the GPU buffer, from uploading & drawing 10k quads.

Shaving that overhead off improves performance on iOS significantly, especially in scenes with a high number of vertex uploads:
| `master` | this PR |
|----------|---------|
| ![IMG_6194](https://github.com/ppy/osu-framework/assets/22781491/f5830571-1c46-4a3d-a38f-e9b5710f9dbe) | ![IMG_6203](https://github.com/ppy/osu-framework/assets/22781491/9815838b-8fad-4b6f-abd9-e8f6acd49926) |
| ![IMG_6195](https://github.com/ppy/osu-framework/assets/22781491/496497db-f244-4ee4-a5f4-3098f78b8c0b) | ![IMG_6204](https://github.com/ppy/osu-framework/assets/22781491/5e2557c2-f688-43ef-b33c-91c1e5c7fc70) |
| ![IMG_6196](https://github.com/ppy/osu-framework/assets/22781491/d7d6e751-2a75-4221-9b37-1de7367ee793) | ![IMG_6205](https://github.com/ppy/osu-framework/assets/22781491/b585eeb2-253a-40ff-98bc-45d8f92d0172) |

Generally, using and writing to a shared buffer directly on Metal is not a bad idea for multiple reasons:
 - It is done for dynamic buffers in Metal's example applications.
 - There's no difference in performance between a shared and a private buffer (no optimisation for GPU access), at least on macOS ([source](https://developer.apple.com/documentation/metal/resource_fundamentals/copying_data_to_a_private_resource?language=objc)):

    ![CleanShot 2023-07-16 at 11 32 10](https://github.com/ppy/osu-framework/assets/22781491/565406c1-46eb-40b7-86e1-36e3693bdac4)


To add this into our implementation appropriately, I've took the time to refactor index buffer storage to become separated from `VeldridVertexBuffer` (and also make more sense), @smoogipoo requesting your review to make sure you're on board with this new structure.

Initially I didn't want to make this change since it keeps on increasing the diversity between backends within the framework project while it still uses Veldrid, but the resultant gain is 100% worth it. Eventually I plan to work on a Metal renderer implementation away from Veldrid, using .NET's API that's shipped with iOS / Mac Catalyst. But for now, I've added this as a separate VBO class for Metal renderers specifically.