@smoogipoo 2023-07-15
Can you also test desktop in the same situation? I tried this a long time ago (before the VBO invalidation work that's been put on the backburner - maybe time to bring that back...) and found similar results that in general we'd get higher FPS by not doing the deduping, but there were certain scenarios where that would fall apart. I can't remember the exact case though, so I hope it's reached by the test cases...

@peppy I also want you to test this on desktop metal, on your x86 PC.