@frenzibyte 2023-07-15
M1 Pro:
| `master` | this PR |
|----------|---------|
| ![CleanShot 2023-07-15 at 20 25 01](https://github.com/ppy/osu-framework/assets/22781491/2701acb6-0a6f-4d2f-bfb4-1b2d3c8ca0c6) | ![CleanShot 2023-07-15 at 20 23 17](https://github.com/ppy/osu-framework/assets/22781491/5010f74b-0a75-4801-a0c0-b25b631d2799) |
| ![CleanShot 2023-07-15 at 20 25 16](https://github.com/ppy/osu-framework/assets/22781491/1243516a-5eb0-4c62-b516-6824e470f60c) | ![CleanShot 2023-07-15 at 20 23 32](https://github.com/ppy/osu-framework/assets/22781491/797e30f4-cb7f-4837-8ca3-900c4cd7b379) |
| ![CleanShot 2023-07-15 at 20 24 01](https://github.com/ppy/osu-framework/assets/22781491/1bee302d-a9f5-42bd-ade7-7a48e66f978f) | ![CleanShot 2023-07-15 at 20 25 21](https://github.com/ppy/osu-framework/assets/22781491/3419a4e4-8123-4284-9741-391510ba7cb8) |