@peppy 2023-07-16
M2 (minor improvements across the board):

| Before | After |
|:------:|:-----:|
| ![2023-07-16 13 24 07@2x](https://github.com/ppy/osu-framework/assets/191335/e2731f0a-3a21-4f79-8762-6f89283bb8de) | ![2023-07-16 13 21 07@2x](https://github.com/ppy/osu-framework/assets/191335/f717a5ec-63cc-4ac3-9d97-411bcc8af652) |
| ![2021-09 13 23 30@2x](https://github.com/ppy/osu-framework/assets/191335/b88e5a35-3f26-46ed-a524-e48cd727016d) | ![2023-07-16 13 20 36@2x](https://github.com/ppy/osu-framework/assets/191335/70014205-e80e-4775-b71a-32433c612478) |
| ![2021-09 13 22 53@2x](https://github.com/ppy/osu-framework/assets/191335/44ab4e35-8ed1-4fd2-a652-c93876dc25be) | ![2023-07-16 13 25 35@2x](https://github.com/ppy/osu-framework/assets/191335/08ad394d-5c43-434d-b2f1-fba3d6b2eeb1) |