@peppy 2023-07-16
macOS / Intel / AMD (considerable improvements across the board!):

| Before | After |
|:------:|:-----:|
| ![osu Framework Tests 2021-09 at 05 41 58](https://github.com/ppy/osu-framework/assets/191335/a6f4c42b-26ef-4f78-b6b3-9b6d034031f9) | ![osu Framework Tests 2023-07-16 at 05 39 49](https://github.com/ppy/osu-framework/assets/191335/a754aa4c-497c-45a7-9f84-dd47a0902994) |
| ![osu Framework Tests 2021-09 at 05 41 19](https://github.com/ppy/osu-framework/assets/191335/deaa8776-6d2f-48fa-a7fa-f5a148f5afea) | ![osu Framework Tests 2023-07-16 at 05 40 42](https://github.com/ppy/osu-framework/assets/191335/3fd70488-19bc-4739-945f-db71c3ea5398) |
| ![osu Framework Tests 2021-09 at 05 41 34](https://github.com/ppy/osu-framework/assets/191335/ca62cd88-8289-4f99-8fd0-2078c9219280) | ![osu Framework Tests 2023-07-16 at 05 40 19](https://github.com/ppy/osu-framework/assets/191335/aad90097-e125-426a-b91f-775e5904564c) |