@peppy 2023-07-16
Windows / Direct X / Nvidia (no difference / within error margins):


| Before | After |
|:------:|:-----:|
|![osu Framework Tests_2023-07-16_16-49-57](https://github.com/ppy/osu-framework/assets/191335/81a7ec34-b9fa-4bc0-bcdf-f36aa15ca573)|![image](https://github.com/ppy/osu-framework/assets/191335/9952901d-3f5e-4005-8143-63a61f0fd43f)|
|![image](https://github.com/ppy/osu-framework/assets/191335/40fc471e-713b-4288-82db-990461724aa9)|![image](https://github.com/ppy/osu-framework/assets/191335/b04aaf5c-bba1-4591-aa31-bab9bdea24b7)|
|![image](https://github.com/ppy/osu-framework/assets/191335/566d24fb-17ab-4e13-bec8-15004227715b)
|![image](https://github.com/ppy/osu-framework/assets/191335/54db36f9-1c57-4cb8-b5aa-582466157c6f)|