@smoogipoo 2023-07-16
As far as I can tell this isn't enabled on D3D, so unless you're making those changes yourself you wouldn't see a difference.