@frenzibyte 2023-07-16
Yeah backends other than Metal are unaffected in this PR.