@smoogipoo 2023-07-21
> And this behaviour is specific to us not 100% guarding against access between CPU & GPU using fences and what not, right

Yes, that's correct.